Transform the path returned by location() to contain native separators
When using a QT installation with a qt.conf configuration file
the prefix will contain forward slashes on windows so that makefiles
generated for nmake will not work as expected. For example moc.exe will
not get called and the build will break at some point where the files
generated by moc are used.