Make QMenu respect the minimum width set
If one sets a minimum width on a QMenu and that size is larger than the
smallest size needed by the large menu item, it ignores the minimum
width and just uses the largest menu item size.

This results in ugly painting artifacts. This currently affects
(at least) the tasks widget in Plasma in KDE4.