Fuzz-compare to define a tolerance with a QCOMPARE
As a customer building a scientific package, specifying a
tolerance "fuzz" factor is imperative to using a unit test
framework.

Some comparisons require a tight tolerance. Others require a loose
tolerance. Numbers might be different based upon casting them to
strings, pushing them onto a QDoubleSpinBox, pushing them through a
binary or text file-write-read step, or deriving a series of numbers
based upon a threaded algorithm.

For these problems, we need a tolerance to use with our calls to
QCOMPARE()