Speed up QFileFormatCollection::indexForFormat
QTextFormatCollection currently has two problems:
 - looking for the index of a QTextFormat is linear, this can take 25% cpu when loading large documents in kword
 - the hash function for QTextFormat is inadequate. Not all values are treated specially. E.g. each QBrush instance has the same hash value at the moment.
These patches speed up loading of a large text document in KWord from 9 to 7 seconds.

This fixes this by using QMultiHash to group the QTextFormat instances by hash and only loop through that list when looking up values. It also improves the hash function for QTextFormat.

Note that the patch is not binary compatible. The patch is binary incompatible in the strictest sense. A variable
 QSet<uint> hashes;
is changed to
 QMultiHash<uint,int> hashes;
It is public in QTextFormatCollection but not used by any other functions.