Fix QPainter::setPen(pen with color but no style) on non-extended.
It was working with a QImage but not with a QPixmap, which is obviously
a bug. In addition, it broke WebCore::GraphicsContext::setPlatformStrokeColor
which does exactly what I put in the unittest: get pen, set color, set pen.
This commit fixes the wrong color in the underline of the links in
http://www.davidfaure.fr/kde/link_underline_color.html in QtWebkit.