Only send the rectangles that really are dirty
This is a small optimization that I left out because I didn't think it
made a difference for most applications. Before this change the paint
event's rect and region are essentially the same.

Well, now I found the difference. If we don't do this, in some
applications that manipulate a lot of text, you get a lot more updates
that normal. This essentially restores the old behavior from Carbon that
region bounding rect is the same as rect, but the actual regions is
potentially smaller.