Dialog loses its modality if user opens a popup within the dialog
When popup was opened SetPointerCapture(..) was always cancelled, but when
it was closed it was not given back to the parent dialog. When modal dialog
was open and user clicked outside the dialog there was no check if the dialog
is modal. Which caused wrong behaviour. Check for the modal dialog was also
added.

Task-number:QTBUG-4810