QNAM HTTP Code: support QNetworkReply::setReadBufferSize() when downloading.
QHttpNetworkReplyPrivate::readBody was always reading as much as possible
from the underlying socket, so even setting the buffer size on the socket
wasn't enough; we need to limit readBody to the available amount of bytes
in the QNetworkReply device, so that it waits until the app has read data.
And then to trigger more reading when that device has room again. And to
flush all that is left when the socket disconnects.

Unit-test included, which tests 4 cases (http/https, with/without buffer size)
and checks the send/receive rates to check that the actual transfer rate
was limited, not just the speed of emission of data to the application.

Task-number: QTBUG-6276