This set of changes is to enable the capability of setting a fragment's visibility. If a fragment is set to invisible the fragment is skipped during the layout in QTextEngine.
This feature has been added due to the need we felt during the implementation of Change Tracking in KWord.

Three new API's has been added
- setVisible and isVisible to QTextFragment
These API's set and test the visibility status of a QTextFragment
- visibleText to QTextBlock
This API returns the visible text of the block i.e only the text that would be considered for layout.

Changes have been made in QTextEngine to skip the invisible fragments during layout.

Note: Fragment's visibility status only affects the Layout. So a find on a QTextDocument would return a cursor position *without* skipping invisible fragments.

Tests have been added to QTextLayout and QTextBlock to test and verify this functionality