Changes: add functionality for dbus auto start to qt
RevBy: TrustMe

Details: This is the first part of an implementation
that introduces handling of the dbus autoStart flag
into qt. This flag determines in the used dbus implementation,
if the receiver of a method call will be started if
it is not running yet, or if it will not be started.

The part from dbus up to QDBusMessage should be fine,
but i do not know how to set this flag from the higher
layers. Ideally this would be generated from qdbusxml2cpp.