QPrinter: Add Current Page print range support
Support the selection of Current Page option in the print dialog by
activating a new QPrintDialog::PrintDialogOption.  If selected the
PrintRange is set to CurrentPage.

It is the responsibility of the application to read the PrintRange and
apply the Current Page selection when rendering the pages.

Unix:
Fully implemented, compiled and tested

QWS:
Fully implemented and compiled, but dialog not tested

Windows:
Fully implemented, but not compiled or tested.

Mac OS X:
Option not supported by Mac OS X print dialog, dialog option ignored as
per existing dialog options.  Not compiled or tested.