QPrinter: Implement Real Number Of Copies as Print Engine Key
Rather than having to do tricks with the print engine using friend
and casting, instead expose the real number of copies in the engine
as a PPK key.

The private function qt_printerRealNumCopies() is no longer used and
may be deleted if it is BC safe to do so.

This is implemented for X11, Windows, and OSX, but has only been
compiled and tested under Linux.