Add eglnullws QScreen driver
This QScreen driver can be used for embedded Qt applications that draw
everything within a single full-screen QGLWidget. It uses underlying
support for a null window system (NullWS) such as is present in PowerVR
systems. It offers performance and stability improvements over the powervr
QScreen driver for applications that do not need full QWS windowing
support.