Fix bug when resetting QSortFilterProxyModel
If multiple proxies are connected to one source model which gets reset,
the order of execution of the sourceModelAboutToBeReset slots could cause
QPersistentModelIndexes to be invalidated in one proxy model before the
same slot is executed in the other proxy model.

Additionally, the persistent indexes in the QSFPM are invalidated before
the persistent indexes in the source model. This patch makes them
invalidated afterward now, which makes more sense.