Fix LeftToRight + setSpacing + setWrapping
Quite a number of assumptions in qlistview assume either wrapping or spacing,
but not both.  This fixes the Flow left to right case.

The TopToBottom case is not affected.  TopToBottom + wrapping still works
TopToBottom + spacing still works, TopToBottom + spacing + wrapping is still
as broken as ever.

Note, should probably be also applied to 4.6 if another patch release is intended.