Overload QGLWidget::renderText to take a QColor as an argument
Existing implementation of QGLWidget::renderText uses
glGetFloatv(GL_CURRENT_COLOR, &color[0]) to determine text color.
New overload passes the color directly.  This is a work-around for an
Intel graphics driver bug which has been persisting since their
8/11/2008 6.14.10.4977 driver for Mobile 4 (and some 965) chipsets on
(at least) XP and Windows 7 64 bit, through at least 3/25/2010
6.14.10.5248.  This bug has affected our customers on (at least)
Panasonic Toughbook and ASUS UL20 devices.  The bug causes
glGetFloatv(GL_CURRENT_COLOR,&color[0]) to return an invalid value,
resulting in a warning from Qt and text to be rendered as black.
Using the workaround enables programs to work properly on devices with
and without the bug.  We implemnted this workaround when the latest
devices delivered by Panasonic would no longer function with a
"rollback" to a driver version that does not have the bug.