Update the state of the keyboard modifiers on drop events.
In the case of drag-n-drop between processes we do not get key events
for the pressed modifiers, so we need to update the state of the modifiers
(using XQueryPointer) to get a correct value of QApplication::keyboardModifiers()
and QDropEvent::keyboardModifiers() at drop time.

This is necessary for fixing KDE bug 178679.