Changes: add patch for artificial emboldening
Details:

Fixes http://bugreports.qt.nokia.com/browse/QTBUG-10596

Patch taken from

https://bugzilla.novell.com/show_bug.cgi?id=374066
https://bugzillafiles.novell.org/attachment.cgi?id=232442

and

http://code.google.com/p/gentoo-taiwan/source/browse/trunk/x11-libs/qt-gui/files/synthetic-bold-4.5.diff?spec=svn298&r=298

and slightly edited.

Moving the two lines concerning the underlining is unrelated to the
emboldening but seems to make sense anyway, the comment "//underline
metrics" seems to be misplaced before the "if (FT_IS_SCALABLE(face)) {|

The hunk:

@@ -789,6 +794,7 @@ QFontEngineFT::Glyph *QFontEngineFT::loadGlyphMetrics(QGlyphSet *set, uint glyph
     }

     FT_GlyphSlot slot = face->glyph;
+    if (embolden) FT_GlyphSlot_Embolden(slot);
     int left  = slot->metrics.horiBearingX;
     int right = slot->metrics.horiBearingX + slot->metrics.width;
     int top    = slot->metrics.horiBearingY;

seems to be logical but it is apparently not used when running
my test program, i.e. synthetic emboldening in my test program
works even without this hunk.