Using Symbian's future font table getter Api
Qt requires raw access to the font tables of used fonts, for a
variety of reasons.

Until Symbian^4, it is/was not possible to access the font tables of
the fonts which are stored in Symbians Font and Bitmap Server.
That's why Qt for Symbian's FontDataBase created an own TFontStore
where it loaded in all installed fonts. While accessing the font
tables via the own TFontStore it still uses the rasterization and
metrics from the FBS, which has public Api for those things.
However, loading all fonts in the own TFontStore for each Qt Gui
process slows down the program startup. Symbian's future font table
setter Api is very welcome and this implemets it's usage.

This patch lets Qt use the new font API on Symbain^4. The font
tables are retrieved via RFontTable.

Task-number: QT-2746
Reviewed-by: Jason Barron