Improve QAccessible for QAccessibleAbstractScrollArea
In QAbstractScrollAreaPrivate::init the created and for the viewPort() used QWidget does setObjectName("qt_scrollarea_viewport"). The QAccessibleAbstractScrollArea then uses that objectName to determinate if the viewPort()'s parentWidget() is a QAbstractScrollArea which is then used for the QAccessibleAbstractScrollArea. This has the disadvantage that by users created and via QAbstractScrollArea::setViewPort set viewPort()'s only work same as the initial viewPort() if they define the same "qt_scrollarea_viewport" objectName. Extend that with a condition that uses w->parentWidget() before falling back to w.