Fix Qt's X11 -geometry position
If the position is left off the -geometry argument Qt is quering the
widget for position, which for most cases will be 0, 0 and calls
main_widget->setGeometry with those values.  As documented in
qt4/doc/html/restoring-geometry.html this is unreliable in X11.  That
document says to use resize and move instead, which also allows the
setup routine to only call resize if a width or height is available
and move if a position is given, leaving the position to the window
manager if it isn't given.

Test cases given to the example moveblocks program.
-geometry 200 -geometry 200x100 -geometry +10+10 -geometry -0-0
-geometry 320x240+10+10
Tested with fvwm and kde window managers.

Task-number: QTBUG-15418
Change-Id: I666eb8ca10c2334e819cf5c424db5df49fcbd0ba
Signed-off-by: David Fries <David@Fries.net>