X11 -geometry test, Display a widget, check the geometry.
These new tests use a helper program to display a main widget and
after a period of time, reads and prints the location of the widget.
This has to be completed in a separate program to test the -geometry
option.  That option is allowed to be set only once and only
the main widget will be affected by it so it requires another program.

The unmodified Qt4.6 fails tst_showGeometry(), with the changes it passes.

This undercovered a race condition with the size and geometry between
the window being created and the window manager placing it.  Sometimes
the geometry is before and sometimes after the move.  The race is
unavoidable, but there needs to be a way to tell Qt to discard the
cached values and update the final values.  The tst_getGeometry() test
uses Qt's geometry values, the tst_showGeometry() for -geometry gets
the values directly from X-windows.

tst_getGeometry() will fail and so the test is set to skip