Fixing bad translations, when Qt has been compiled with Q_BACKINGSTORE_SUBSURFACES ( f.e when building OpenGL/Qt-Embedded )
When painting to the backingstore or blitting from the backing store
to the surface coordinates depend on the offset of the window frame
and the widget position relative to the top level window.
This patch fixes a couple of wrong translations in the
implementation, that is enabled when Q_BACKINGSTORE_SUBSURFACES is on.

These changes fix all wrong translations I noticed in my application,
beside a problem with toplevel OpenGL windows with a frame. The reason
for this problem can be found at qpaintengine_opengl.cpp line 1367.
Unfortunately this one can't be fixed easily, because the offset of
the frame is not known at this place. The comment at this line
shows, that the author was aware of this problem - but obviously got
forgotten.

 ---[ Fields ]-----------------[ uncomment and edit as applicable ]---|