Correct gravity for negative -geometry positions.
The window gravity was always set to NorthEast or NorthWest, according
to the application right to left setting.  The window gravity defines
the direction the window manager will move the window when adding
borders.  Specifying +0+0 and NorthEast will shift the window right
and down so the top left pixel of the window border will be at 0,0.
See man XSizeHints(3) for details.  Without modifying the gravity -0-0
will exactly put the window at the bottom right of the screen, but
then the window manager shifts it down and right (or left for
NorthWest) partly off the screen, or depending on the window manager
it will shift it back on the screen causing all valus less than or
equal to the border size to land in the same location.

This change sets the window gravity if any of the X position values
are negative in
QApplicationPrivate::applyX11SpecificCommandLineArguments
and retains the current gravity when moving and resizing afterwords.