Export qt_directfb_.* functions in plugin as well
These functions can be a big help in debugging rendering errors in an
application and can even be handy to use for real functionality. While
it requires the app to explicitly link to the plugin (when compiling
with -plugin-gfx-directfb this is still much better than not exporting
them).

Since this was the only use-case for QT_DIRECTFB_PLUGIN it's probably
better to take that define out.