QTBUG-12116: Fixed qargb8565 constructor taking quint32 as argument. Corrected encoding into 16bit of RGB numbers. This fixes issues when using transparancy in combination with 16bit framebuffers.
Internally the qargb8565 uses an array of bytes (3bytes) to represent the 24bit color.
Byte 0 is used to stored the alpha channel.
Bytes 1 and 2 are used to store the RGB values in 565 encoding, starting with red, then green and finally bleu.
This is now corrected since byte 1 and 2 were swapped. This constructor is used when a 32bit color is converted into a 24bit color (e.g. a PNG containing alpha channel and a 16bit framebuffer). In this case, it is not painting correctly.