Fixed crash that happens after destroying Phonon::QT7::VideoWidget
QTMovieView needs to perform some cleanup in its own implementation of
viewWillMoveToWindow.  When SharedQTMovieView overrides the function
without calling QTMovieView's viewWillMoveToWindow, the cleanup is not
performed.

The cleanup involves removing itself as an observer from the
NSNotificationCenter.  If QTMovieView has been deallocated without
performing the cleanup, the NSNotificationCenter will continue to send
notifications to QTMovieView even though it have been deallocated, which
causes a crash.