use qBinaryFind instead of bsearch
depending on platform, qBinaryFind might be up to 3 times faster than bsearch
or a bit bit slower at the worst case; so why don't simply use qBinaryFind instead of bsearch and be an optimist? ;)

also we know that all color names are in the lower case already and thus
we can avoid lowercasing of them again and again, and again;
the name to find needs to be lowercased separately instead.

some color-by-name-lookup numbers:
before:
 0.000737 msecs per iteration (total: 3,684, iterations: 5000000)
after:
 0.000154 msecs per iteration (total: 771, iterations: 5000000)