fix include path
In many situations building succeeds without explicity including
this dir in the include path, probably due to a side-effect of
calling uic on the form in this project. However, it is known
to fail at least on mingw-cross-env. Anyway, it makes sense
to explcitly add this dir to the include path since a needed
header resides here.