Add comctl32 v6 assembly dependency to applications only (not libraries)
This improves the patch from merge request #2526 to add the comctl32 v6 assembly dependency to applications only, as it was being done previously (although only when /SUBSYSTEM:WINDOWS, not when /SUBSYSTEM:CONSOLE). Other existing makespecs are not affected as the $$QMAKE_LFLAGS_EXE variable evaluates to empty string when not defined.